requirements.txt
asgiref==3.8.1
Django==5.1.6
django-allauth==65.4.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
